IdeaInspiration/Sources/Signals/News/NewsApi/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
newsapi-python
sqlite-utils>=3.35
click>=8.1.7

# Development dependencies (install with: pip install -r requirements.txt[dev])
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
